requirements.txt
streamlit
requests
python-dotenv
deepgram-sdk
langchain-groq
langchain-core
deep-translator
openai-whisper
ffmpeg
langchain
